#277 [refactor] SingleEventHandler 적용
